Pushing FFT script and play audio from frequency script
the frequency script is a mock up created by chatGPT but shows how to play audio in unity from an array of values.

FFT takes in an array of complex number and outputs an array of complex numbers.

maybe use default values of 1 second sample rate and 4401 audio frequency in the script to generate the frequency for playing ... ???